Fix for storageNode params and CDNNode params unmarshal (#56)

+++ contract/pkg/mock/ddc_bucket_contract_mock.go
@@ -48,7 +48,7 @@ type (
 
 	CDNNodeParams struct {
 		Url       string `json:"url"`
-		Size      uint8  `json:"size"`
+		Size      int    `json:"size"`
 		Location  string `json:"location"`
 		PublicKey string `json:"publicKey"`
 	}
